Refactor `TypeUtil` into `ClassUtil`
* Refactored the `TypeUtil` class into the `ClassUtil` as both
  serve similar jobs
* Fixed more misc. warnings
