+++ src/main/java/org/assertj/assertions/generator/description/FieldDescription.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
  *
@@ -14,7 +14,6 @@ package org.assertj.assertions.generator.description;
 
 import com.google.common.reflect.TypeToken;
 import org.assertj.assertions.generator.util.ClassUtil;
-import org.assertj.assertions.generator.util.TypeUtil;
 
 import java.lang.reflect.Field;
 
@@ -62,7 +61,7 @@ public class FieldDescription extends DataDescription implements Comparable<Fiel
 
   @Override
   public boolean isPredicate() {
-    return TypeUtil.isBoolean(valueType);
+    return ClassUtil.isBoolean(valueType);
   }
 
   @Override
